+++ src/test/java/com/uber/h3core/benchmarking/PolyfillBenchmark.java
@@ -17,7 +17,7 @@ package com.uber.h3core.benchmarking;
 
 import com.google.common.collect.ImmutableList;
 import com.uber.h3core.H3Core;
-import com.uber.h3core.util.Vector2D;
+import com.uber.h3core.util.GeoCoord;
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
 import org.openjdk.jmh.annotations.Mode;
@@ -40,12 +40,12 @@ public class PolyfillBenchmark {
     public List<Long> benchmarkPolyfill() {
         return BenchmarkState.h3Core.polyfill(
                 ImmutableList.of(
-                        new Vector2D(37.813318999983238, -122.4089866999972145),
-                        new Vector2D(37.7866302000007224, -122.3805436999997056),
-                        new Vector2D(37.7198061999978478, -122.3544736999993603),
-                        new Vector2D(37.7076131999975672, -122.5123436999983966),
-                        new Vector2D(37.7835871999971715, -122.5247187000021967),
-                        new Vector2D(37.8151571999998453, -122.4798767000009008)
+                        new GeoCoord(37.813318999983238, -122.4089866999972145),
+                        new GeoCoord(37.7866302000007224, -122.3805436999997056),
+                        new GeoCoord(37.7198061999978478, -122.3544736999993603),
+                        new GeoCoord(37.7076131999975672, -122.5123436999983966),
+                        new GeoCoord(37.7835871999971715, -122.5247187000021967),
+                        new GeoCoord(37.8151571999998453, -122.4798767000009008)
                 ), null, 9
         );
     }
@@ -55,18 +55,18 @@ public class PolyfillBenchmark {
     public List<Long> benchmarkPolyfillWithHole() {
         return BenchmarkState.h3Core.polyfill(
                 ImmutableList.of(
-                        new Vector2D(37.813318999983238, -122.4089866999972145),
-                        new Vector2D(37.7866302000007224, -122.3805436999997056),
-                        new Vector2D(37.7198061999978478, -122.3544736999993603),
-                        new Vector2D(37.7076131999975672, -122.5123436999983966),
-                        new Vector2D(37.7835871999971715, -122.5247187000021967),
-                        new Vector2D(37.8151571999998453, -122.4798767000009008)
+                        new GeoCoord(37.813318999983238, -122.4089866999972145),
+                        new GeoCoord(37.7866302000007224, -122.3805436999997056),
+                        new GeoCoord(37.7198061999978478, -122.3544736999993603),
+                        new GeoCoord(37.7076131999975672, -122.5123436999983966),
+                        new GeoCoord(37.7835871999971715, -122.5247187000021967),
+                        new GeoCoord(37.8151571999998453, -122.4798767000009008)
                 ),
                 ImmutableList.of(
                         ImmutableList.of(
-                                new Vector2D(37.7869802, -122.4471197),
-                                new Vector2D(37.7664102, -122.4590777),
-                                new Vector2D(37.7710682, -122.4137097)
+                                new GeoCoord(37.7869802, -122.4471197),
+                                new GeoCoord(37.7664102, -122.4590777),
+                                new GeoCoord(37.7710682, -122.4137097)
                         )
                 ),
                 9
@@ -78,23 +78,23 @@ public class PolyfillBenchmark {
     public List<Long> benchmarkPolyfillWithTwoHoles() {
         return BenchmarkState.h3Core.polyfill(
                 ImmutableList.of(
-                        new Vector2D(37.813318999983238, -122.4089866999972145),
-                        new Vector2D(37.7866302000007224, -122.3805436999997056),
-                        new Vector2D(37.7198061999978478, -122.3544736999993603),
-                        new Vector2D(37.7076131999975672, -122.5123436999983966),
-                        new Vector2D(37.7835871999971715, -122.5247187000021967),
-                        new Vector2D(37.8151571999998453, -122.4798767000009008)
+                        new GeoCoord(37.813318999983238, -122.4089866999972145),
+                        new GeoCoord(37.7866302000007224, -122.3805436999997056),
+                        new GeoCoord(37.7198061999978478, -122.3544736999993603),
+                        new GeoCoord(37.7076131999975672, -122.5123436999983966),
+                        new GeoCoord(37.7835871999971715, -122.5247187000021967),
+                        new GeoCoord(37.8151571999998453, -122.4798767000009008)
                 ),
                 ImmutableList.of(
                         ImmutableList.of(
-                                new Vector2D(37.7869802, -122.4471197),
-                                new Vector2D(37.7664102, -122.4590777),
-                                new Vector2D(37.7710682, -122.4137097)
+                                new GeoCoord(37.7869802, -122.4471197),
+                                new GeoCoord(37.7664102, -122.4590777),
+                                new GeoCoord(37.7710682, -122.4137097)
                         ),
                         ImmutableList.of(
-                                new Vector2D(37.747976, -122.490025),
-                                new Vector2D(37.731550, -122.503758),
-                                new Vector2D(37.725440, -122.452603)
+                                new GeoCoord(37.747976, -122.490025),
+                                new GeoCoord(37.731550, -122.503758),
+                                new GeoCoord(37.725440, -122.452603)
                         )
                 ),
                 9